resume-nlp-backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
blis==1.3.0
catalogue==2.0.10
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
cloudpathlib==0.21.1
colorama==0.4.6
confection==0.1.5
cymem==2.0.11
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl
exceptiongroup==1.3.0
fastapi==0.115.14
h11==0.16.0
idna==3.10
Jinja2==3.1.6
joblib==1.5.1
langcodes==3.5.0
language_data==1.3.0
language_tool_python==2.9.4
lxml==6.0.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
murmurhash==1.0.13
numpy==2.2.6
packaging==25.0
preshed==3.0.10
psutil==7.0.0
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
PyPDF2==3.0.1
python-docx==1.2.0
python-multipart==0.0.20
requests==2.32.4
rich==14.0.0
scikit-learn==1.7.0
scipy==1.15.3
shellingham==1.5.4
smart-open==7.1.0
sniffio==1.3.1
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
starlette==0.46.2
thinc==8.3.6
threadpoolctl==3.6.0
toml==0.10.2
tqdm==4.67.1
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.5.0
uvicorn==0.35.0
wasabi==1.1.3
weasel==0.4.1
wrapt==1.17.2
